add areas to filter in templates, move more functions to helpers when shared
